refactor(darwin): rename backend package darwin -> vfkit
Name the macOS backend after its driver (vfkit) to mirror the libvirt
backend, which is named for its driver rather than the OS. The registered
backend Name was already "vfkit"; this aligns the Go package/directory and
the comments that referred to the macOS backend as "darwin".

OS-meaning constructs are deliberately left as-is: //go:build darwin
constraints, _darwin.go file suffixes, and comments that genuinely refer
to the macOS operating system.

Also fix a Linux assumption leaking into cross-platform code: importcmd
hardcoded a ".qcow2" base-image extension instead of routing through
config.UserBaseImageName, which resolves to ".raw" on macOS.
